Move menu generator to wcc namespace
  so it can be invoked with `rails generate wcc:menu`
