+++ lib/generators/wcc/USAGE
