+++ lib/generators/wcc/menu_generator.rb
