+++ lib/generators/wcc/templates/Procfile
