+++ lib/generators/wcc/templates/contentful
