+++ lib/generators/wcc/templates/generated_add_menus.ts
