+++ lib/generators/wcc/templates/release
