+++ lib/generators/wcc/templates/wcc_contentful.rb
